+++ src/index.js
@@ -1,1 +1,2 @@
+export BodyContainer from "./components/BodyContainer"
 export joinUri from "./_utils/join-uri"